requirements.txt
fastapi
uvicorn
sqlmodel
asyncpg
alembic
pydantic
httpx
python-dotenv
celery[redis]
cryptography
bip-utils
web3
passlib[bcrypt]
python-jose[cryptography]
slowapi
pydantic[email]
nest_asyncio
asgiref
python-multipart 